Add user ubuntu in dev container
The user ubuntu is only introduced in the Docker image starting with version 24.04. The jammy LTS 22.04 does not yet have it. See https://bugs.launchpad.net/cloud-images/+bug/2005129 for details.

Apart from being correct to add the user when it does not exist, this also represents a workaround for issues with dev container tools who do not add the user automatically when remoteUser is set. JetBrains dev container integration in CLion is affected by this. See for example https://youtrack.jetbrains.com/issue/IJPL-160538/DevContainers.-Fails-to-create-a-devcontainer-with-not-existing-remoteUser-with-unclear-message-No-Linux-Installation-Found-on