Introduce AOD mode without dimming for stationary connected devices
Add a value of "0ff" to the display timeout settings.
This disables the dimming and color reduction associated with AOD mode, meaning the screen will look the same at all time. The 2 Hz idle frame rate is kept, as that is still satisfactory.
This connected AOD mode can be desirable for usage in a stationary setting such as a weather station or desk clock because not all apps are designed with the color reduction in mind and the brightness is very low. Higher power usage is not a concern when always connected to a charger.

UI in the display settings is not yet perfect in my opinion. A better option might be to keep the duration like this and replace the AOD checkbox with a "Timeout action" section with two radiobuttons:
- Turn off
- Dim
 Turn off would be like the current non-AOD behaviour where the screen is turned off after the display timeout.
Dim would be the behaviour of the current AOD mode, where brightness is reduced sharply and a power saving color + frame rate reduced mode is entered.
The newly introduced "connected" AOD mode where the display is kept completely on would be selected with the 0 s display timeout.
I think that alternative is a bit better to understand than the current setup.